+++ flutter-hms-push/android/src/main/java/com/huawei/hms/flutter/push/utils/BundleUtils.java
@@ -1,11 +1,11 @@
 /*
-Copyright 2020. Huawei Technologies Co., Ltd. All rights reserved.
+    Copyright 2020. Huawei Technologies Co., Ltd. All rights reserved.
 
-    Licensed under the Apache License, Version 2.0 (the "License");
+    Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
     You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+        https://www.apache.org/licenses/LICENSE-2.0
 
     Unless required by applicable law or agreed to in writing, software
     distributed under the License is distributed on an "AS IS" BASIS,
@@ -26,6 +26,11 @@ import org.json.JSONObject;
 import java.util.Set;
 
 public class BundleUtils {
+
+    private BundleUtils() {
+        throw new IllegalStateException("Utility class");
+    }
+
     public static String get(Bundle bundle, String key) {
 
         return bundle != null ? bundle.getString(key) : null;
@@ -90,32 +95,32 @@ public class BundleUtils {
     }
 
     public static String convertJSON(Bundle bundle) {
-        try {
-            JSONObject json = convertJSONObject(bundle);
-            return json.toString();
-        } catch (JSONException e) {
-            Log.d("BundleUtils", "Error occurred while converting bundle to JSON String");
-            return "";
-        }
+
+        JSONObject json = convertJSONObject(bundle);
+        return json.toString();
+
     }
 
-    private static JSONObject convertJSONObject(Bundle bundle) throws JSONException {
-
-        JSONObject json = new JSONObject();
-        if (bundle != null) {
-            Set<String> keys = bundle.keySet();
-            for (String key : keys) {
-                Object value = bundle.get(key);
-                if (value instanceof Bundle) {
-                    json.put(key, convertJSONObject((Bundle) value));
-                } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-                    json.put(key, JSONObject.wrap(value));
-                } else {
-                    json.put(key, value);
+    static public JSONObject convertJSONObject(Bundle bundle) {
+        try {
+            JSONObject json = new JSONObject();
+            if (bundle != null) {
+                Set<String> keys = bundle.keySet();
+                for (String key : keys) {
+                    Object value = bundle.get(key);
+                    if (value instanceof Bundle) {
+                        json.put(key, convertJSONObject((Bundle) value));
+                    } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+                        json.put(key, JSONObject.wrap(value));
+                    } else {
+                        json.put(key, value);
+                    }
                 }
             }
+            return json;
+        } catch (JSONException e) {
+            return null;
         }
-        return json;
     }
 
 }