+++ flutter-hms-push/lib/constants/channel.dart
@@ -1,11 +1,11 @@
 /*
-Copyright 2020. Huawei Technologies Co., Ltd. All rights reserved.
+    Copyright 2020. Huawei Technologies Co., Ltd. All rights reserved.
 
-    Licensed under the Apache License, Version 2.0 (the "License");
+    Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
     You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+        https://www.apache.org/licenses/LICENSE-2.0
 
     Unless required by applicable law or agreed to in writing, software
     distributed under the License is distributed on an "AS IS" BASIS,
@@ -27,3 +27,4 @@ const String NOTIFICATION_OPEN_CHANNEL =
 const String LOCAL_NOTIFICATION_CLICK_CHANNEL =
     "com.huawei.flutter.push/local_notification_click";
 const String METHOD_CHANNEL = "com.huawei.flutter.push/method";
+const String BACKGROUND_MESSAGE_CHANNEL = "com.huawei.flutter.push/background";