+++ flutter-hms-push/lib/push.dart
@@ -1,11 +1,11 @@
 /*
-Copyright 2020. Huawei Technologies Co., Ltd. All rights reserved.
+    Copyright 2020. Huawei Technologies Co., Ltd. All rights reserved.
 
-    Licensed under the Apache License, Version 2.0 (the "License");
+    Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
     You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+        https://www.apache.org/licenses/LICENSE-2.0
 
     Unless required by applicable law or agreed to in writing, software
     distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,7 +16,9 @@ Copyright 2020. Huawei Technologies Co., Ltd. All rights reserved.
 
 import 'dart:async';
 import 'dart:convert';
+import 'dart:ui';
 
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:huawei_push/model/remote_message.dart';
 
@@ -26,7 +28,7 @@ import 'constants/Code.dart';
 
 class Push {
   /// Constant Method and Event Channel definitions for communicating with native platform
-  static const MethodChannel mChannel =
+  static const MethodChannel methodChannel =
       const MethodChannel(Channel.METHOD_CHANNEL);
   static const EventChannel tokenEventChannel =
       EventChannel(Channel.TOKEN_CHANNEL);
@@ -45,7 +47,7 @@ class Push {
   ///
   /// Returns the corresponding Push SDK Result Code's Description
   static Future<String> turnOnPush() async {
-    final String result = await mChannel.invokeMethod(Method.turnOnPush);
+    final String result = await methodChannel.invokeMethod(Method.turnOnPush);
     return Code[result];
   }
 
@@ -53,31 +55,31 @@ class Push {
   ///
   /// Returns the corresponding Push SDK Result Code's Description
   static Future<String> turnOffPush() async {
-    final String result = await mChannel.invokeMethod(Method.turnOffPush);
+    final String result = await methodChannel.invokeMethod(Method.turnOffPush);
     return Code[result];
   }
 
   /// Obtains an AAID in synchronous mode of Push SDK
   static Future<String> getId() async {
-    final String result = await mChannel.invokeMethod(Method.getId);
+    final String result = await methodChannel.invokeMethod(Method.getId);
     return result;
   }
 
   /// Obtains an AAID in asynchronous mode of Push SDK
   static Future<String> getAAID() async {
-    final String result = await mChannel.invokeMethod(Method.getAAID);
+    final String result = await methodChannel.invokeMethod(Method.getAAID);
     return result;
   }
 
   /// Obtains the Application ID from the *agconnect-services.json* file
   static Future<String> getAppId() async {
-    final String result = await mChannel.invokeMethod(Method.getAppId);
+    final String result = await methodChannel.invokeMethod(Method.getAppId);
     return result;
   }
 
   /// Obtains an ODID
   static Future<String> getOdid() async {
-    final String result = await mChannel.invokeMethod(Method.getOdid);
+    final String result = await methodChannel.invokeMethod(Method.getOdid);
     return result;
   }
 
@@ -85,19 +87,23 @@ class Push {
   ///
   /// To obtain the push token one must listen the token stream. Token Stream
   /// can be obtained by [getTokenStream] method
-  static Future<void> getToken() async {
-    mChannel.invokeMethod<String>(Method.getToken);
+  static Future<void> getToken(String scope) async {
+    Map<String, dynamic> args = <String, dynamic>{};
+    args.putIfAbsent("scope", () => scope);
+
+    methodChannel.invokeMethod<String>(Method.getToken, args);
   }
 
   /// Obtains the generation timestamp of an AAID.
   static Future<String> getCreationTime() async {
-    final String result = await mChannel.invokeMethod(Method.getCreationTime);
+    final String result =
+        await methodChannel.invokeMethod(Method.getCreationTime);
     return result;
   }
 
   /// Deletes a local AAID and its generation timestamp.
   static Future<String> deleteAAID() async {
-    final String result = await mChannel.invokeMethod(Method.deleteAAID);
+    final String result = await methodChannel.invokeMethod(Method.deleteAAID);
     return Code[result];
   }
 
@@ -106,8 +112,11 @@ class Push {
   /// Native Push SDK doesn't allow this operation on main thread which doesn't
   /// comply with method channels’ UI thread requirement of Flutter, so the result
   /// string and errors will be returned on TokenEventChannel's [onError] callback.
-  static Future<void> deleteToken() async {
-    mChannel.invokeMethod(Method.deleteToken);
+  static Future<void> deleteToken(String scope) async {
+    Map<String, dynamic> args = <String, dynamic>{};
+    args.putIfAbsent("scope", () => scope);
+
+    methodChannel.invokeMethod(Method.deleteToken, args);
   }
 
   /// Getter for TokenEventChannel's Stream, which emits the requested push token
@@ -141,7 +150,7 @@ class Push {
   static Future<String> sendRemoteMessage(
       RemoteMessageBuilder remoteMsg) async {
     final String result =
-        await mChannel.invokeMethod(Method.send, remoteMsg.toMap());
+        await methodChannel.invokeMethod(Method.send, remoteMsg.toMap());
     return Code[result];
   }
 
@@ -159,7 +168,8 @@ class Push {
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent("topic", () => topic);
 
-    final String result = await mChannel.invokeMethod(Method.subscribe, args);
+    final String result =
+        await methodChannel.invokeMethod(Method.subscribe, args);
     return Code[result];
   }
 
@@ -168,7 +178,8 @@ class Push {
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent("topic", () => topic);
 
-    final String result = await mChannel.invokeMethod(Method.unsubscribe, args);
+    final String result =
+        await methodChannel.invokeMethod(Method.unsubscribe, args);
     return Code[result];
   }
 
@@ -188,20 +199,21 @@ class Push {
     args.putIfAbsent("enabled", () => enabled);
 
     final String result =
-        await mChannel.invokeMethod(Method.setAutoInitEnabled, args);
+        await methodChannel.invokeMethod(Method.setAutoInitEnabled, args);
     return Code[result];
   }
 
   /// Checks whether automatic initialization is enabled.
   static Future<bool> isAutoInitEnabled() async {
-    final bool result = await mChannel.invokeMethod(Method.isAutoInitEnabled);
+    final bool result =
+        await methodChannel.invokeMethod(Method.isAutoInitEnabled);
     return result;
   }
 
   /// Gets the values of the **agconnect-services.json** file
   static Future<String> getAgConnectValues() async {
     final String result =
-        await mChannel.invokeMethod<String>(Method.getAgConnectValues);
+        await methodChannel.invokeMethod<String>(Method.getAgConnectValues);
     return result;
   }
 
@@ -209,7 +221,7 @@ class Push {
   static Future<Null> showToast(String msg) async {
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent("msg", () => msg);
-    await mChannel.invokeMethod(Method.showToast, args);
+    await methodChannel.invokeMethod(Method.showToast, args);
     return null;
   }
 
@@ -217,7 +229,8 @@ class Push {
   /// Updates itself with the latest custom intent value if another Remote Notification Message
   /// with custom intent is tapped.
   static Future<String> getInitialIntent() async {
-    final String result = await mChannel.invokeMethod(Method.getInitialIntent);
+    final String result =
+        await methodChannel.invokeMethod(Method.getInitialIntent);
     return result;
   }
 
@@ -231,12 +244,12 @@ class Push {
           .receiveBroadcastStream()
           .cast<String>();
 
-  /// Returns a Remote Message object representing the notification which opened the app.
-  static Future<RemoteMessage> getInitialNotification() async {
-    final result = await mChannel.invokeMethod(Method.getInitialNotification);
+  /// Returns a Map object representing the notification which opened the app.
+  static Future<dynamic> getInitialNotification() async {
+    final result =
+        await methodChannel.invokeMethod(Method.getInitialNotification);
     if (result != null) {
-      Map<String, dynamic> resultMap = Map.from(result);
-      return RemoteMessage.fromMap(resultMap);
+      return result;
     }
     return null;
   }
@@ -244,10 +257,11 @@ class Push {
   /// Getter for the NotificationOpenEventChannel's Stream.
   ///
   /// This stream emits a remote message object representing the clicked notification which opened the app.
-  static Stream<RemoteMessage> get onNotificationOpenedApp =>
+  static Stream<dynamic> get onNotificationOpenedApp =>
       notificationOpenEventChannel
           .receiveBroadcastStream()
-          .map((event) => RemoteMessage.fromMap(json.decode(event)));
+          .map((event) => json.decode(event))
+          .cast<Map<String, dynamic>>();
 
   /// Getter for the LocalNotificationClickEventChannel's Stream
   ///
@@ -262,7 +276,7 @@ class Push {
   /// Pushes a local notification instantly
   static Future<Map<String, dynamic>> localNotification(
       Map<String, dynamic> localNotification) async {
-    final result = await mChannel.invokeMethod(
+    final result = await methodChannel.invokeMethod(
         Method.localNotification, localNotification);
     Map<String, dynamic> resultMap = json.decode(result);
     return resultMap;
@@ -271,7 +285,7 @@ class Push {
   /// Schedules a local notification
   static Future<Map<String, dynamic>> localNotificationSchedule(
       Map<String, dynamic> localNotification) async {
-    final result = await mChannel.invokeMethod(
+    final result = await methodChannel.invokeMethod(
         Method.localNotificationSchedule, localNotification);
     Map<String, dynamic> resultMap = json.decode(result);
     return resultMap;
@@ -280,7 +294,7 @@ class Push {
   /// Returns the list of all active notifications
   static Future<List<Map<String, dynamic>>> getNotifications() async {
     final List<dynamic> notifications =
-        await mChannel.invokeMethod(Method.getNotifications);
+        await methodChannel.invokeMethod(Method.getNotifications);
     List<Map<String, dynamic>> result = [];
     notifications.forEach((element) {
       result.add(json.decode(element));
@@ -291,7 +305,7 @@ class Push {
   /// Returns a list of all pending scheduled notifications
   static Future<List<Map<String, dynamic>>> getScheduledNotifications() async {
     final List<dynamic> scheduledNotifications =
-        await mChannel.invokeMethod(Method.getScheduledNotifications);
+        await methodChannel.invokeMethod(Method.getScheduledNotifications);
     List<Map<String, dynamic>> result = [];
     scheduledNotifications.forEach((element) {
       result.add(json.decode(element));
@@ -301,7 +315,7 @@ class Push {
 
   /// Returns the list of all notification channels.
   static Future<List<String>> getChannels() async {
-    final List result = await mChannel.invokeMethod(Method.getChannels);
+    final List result = await methodChannel.invokeMethod(Method.getChannels);
     List<String> strList = result.cast<String>();
     return strList;
   }
@@ -309,62 +323,98 @@ class Push {
   /// Deletes a notification channel with the specified ID
   static Future<String> deleteChannel(String channelId) async {
     final String result =
-        await mChannel.invokeMethod(Method.deleteChannel, channelId);
+        await methodChannel.invokeMethod(Method.deleteChannel, channelId);
     return Code[result];
   }
 
   /// Checks if the channel with the specified ID exists
   static Future<bool> channelExists(String channelId) async {
-    final result = await mChannel.invokeMethod(Method.channelExists, channelId);
+    final result =
+        await methodChannel.invokeMethod(Method.channelExists, channelId);
     return result;
   }
 
   /// Checks if the notification channel with the specified ID is blocked
   static Future<bool> channelBlocked(String channelId) async {
     final result =
-        await mChannel.invokeMethod(Method.channelBlocked, channelId);
+        await methodChannel.invokeMethod(Method.channelBlocked, channelId);
     return result;
   }
 
   /// Cancels all pending notifications registered in the notification manager
   static Future<void> cancelNotifications() async {
-    mChannel.invokeMethod(Method.cancelNotifications);
+    methodChannel.invokeMethod(Method.cancelNotifications);
   }
 
   /// Cancels all pending scheduled notifications and the ones registered in the
   /// notification manager
   static Future<void> cancelAllNotifications() async {
-    mChannel.invokeMethod(Method.cancelAllNotifications);
+    methodChannel.invokeMethod(Method.cancelAllNotifications);
   }
 
   /// Cancels all pending scheduled notifications
   static Future<void> cancelScheduledNotifications() async {
-    mChannel.invokeMethod(Method.cancelScheduledNotifications);
+    methodChannel.invokeMethod(Method.cancelScheduledNotifications);
   }
 
   /// Cancels all active notifications with the specified tag
   static Future<void> cancelNotificationsWithTag(String tag) async {
-    mChannel.invokeMethod(Method.cancelNotificationsWithTag, tag);
+    methodChannel.invokeMethod(Method.cancelNotificationsWithTag, tag);
   }
 
   /// Cancels all notification with the list of specified ID
   static Future<void> cancelNotificationsWithId(List<int> ids) async {
-    mChannel.invokeMethod(Method.cancelNotificationsWithId, ids);
+    methodChannel.invokeMethod(Method.cancelNotificationsWithId, ids);
   }
 
   /// Cancels all notifications with specified ID-Tag Map
   static Future<void> cancelNotificationsWithIdTag(
       Map<int, String> idTags) async {
-    mChannel.invokeMethod(Method.cancelNotificationsWithIdTag, idTags);
+    methodChannel.invokeMethod(Method.cancelNotificationsWithIdTag, idTags);
   }
 
   /// Enables HMS Plugin Method Analytics
   static Future<void> enableLogger() async {
-    mChannel.invokeMethod(Method.enableLogger);
+    methodChannel.invokeMethod(Method.enableLogger);
   }
 
   /// Disables HMS Plugin Method Analytics
   static Future<void> disableLogger() async {
-    mChannel.invokeMethod(Method.disableLogger);
+    methodChannel.invokeMethod(Method.disableLogger);
   }
+
+  static Future<bool> registerBackgroundMessageHandler(
+      void Function(RemoteMessage remoteMessage) callback) {
+    int rawHandle =
+        PluginUtilities.getCallbackHandle(callbackDispatcher).toRawHandle();
+    print("rawHandle $rawHandle");
+    int rawCallback = PluginUtilities.getCallbackHandle(callback).toRawHandle();
+    print("rawCallback $rawCallback");
+    return methodChannel.invokeMethod('registerBackgroundMessageHandler', {
+      "rawHandle": rawHandle,
+      "rawCallback": rawCallback
+    }).then((result) => result == 1);
+  }
+
+  static Future<bool> removeBackgroundMessageHandler() {
+    return methodChannel
+        .invokeMethod('removeBackgroundMessageHandler')
+        .then((result) => result == 1);
+  }
+}
+
+void callbackDispatcher() {
+  const MethodChannel backgroundMessageChannel =
+      MethodChannel(Channel.BACKGROUND_MESSAGE_CHANNEL);
+  WidgetsFlutterBinding.ensureInitialized();
+
+  backgroundMessageChannel.setMethodCallHandler((MethodCall call) async {
+    final Map<String, dynamic> args =
+        Map<String, dynamic>.from(call.arguments[1]);
+    RemoteMessage remoteMessage = RemoteMessage.fromMap(args);
+    final Function rawHandler = PluginUtilities.getCallbackFromHandle(
+        CallbackHandle.fromRawHandle(call.arguments[0]));
+
+    rawHandler(remoteMessage);
+  });
 }